requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pymongo==4.15.3
langchain==0.3.0
langchain-openai==0.2.0

